[AN] refactor: 도메인 예외 클래스 패키지 구조 변경 #718
*   [AN] refactor: DomainException, AuthException을 domain.exception 패키지로 이동
