+++ android/app/src/main/java/com/onair/hearit/domain/exception/DomainException.kt
@@ -1,4 +1,4 @@
-package com.onair.hearit.domain
+package com.onair.hearit.domain.exception
 
 sealed class DomainException : Exception() {
     data object NetworkConnection : DomainException()
